+++ client/package.json
@@ -13,7 +13,7 @@
   },
   "scripts": {
     "build": "turbo run build --cache-dir=.turbo --no-update-notifier",
-    "dev": "turbo run dev --parallel --no-update-notifier",
+    "dev": "turbo run dev --filter=!@instantdb/mcp --parallel --no-update-notifier",
     "test": "turbo run test:ci --no-update-notifier",
     "bench": "turbo run bench:ci --no-update-notifier",
     "format": "prettier --write --ignore-path ../.gitignore --config ./.prettierrc \"**/*.{ts,tsx,js,jsx,json,md}\"",
